chore(types): add type imports to component type generation
this commit passes type import data to the function that helps
generate type information for a component. this commit is split off from
the rest of the this effort to allow for it to be shared amongst
multiple PRs.

STENCIL-27: duplicate named types in public aps do not get assigned
correctly in components.d.ts