refactor(types): update variable names, add jsdoc
this commit refactors the `update-import-refs` helper module by adding
jsdoc and renaming existing variables in the name of improving clarity.
this module utilizes multiple data structures that cover similar aspects
of the type resolution process, but can be difficult to keep a mental
model of. the intent of this refactor is to clarify the intent/purpose
of this code.

STENCIL-27: duplicate named types in public apis do not get assigned
correctly in components.d.ts